Exceptions added

+++ CA2/src/java/facade/PersonFacade.java
@@ -7,10 +7,10 @@ package facade;
 
 import interfaces.IFPersonFacade;
 import entity.Person;
+import exceptions.PersonNotFoundException;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
-import javax.persistence.Query;
 
 /**
  *
@@ -29,10 +29,13 @@ public class PersonFacade implements IFPersonFacade {
     }
 
     @Override
-    public Person getPerson(int id) {
+    public Person getPerson(int id) throws PersonNotFoundException {
         EntityManager em = getEntityManager();
         try {
             Person p = em.find(Person.class, id);
+            if (p == null) {
+                throw new PersonNotFoundException("No Person found with provided id");
+            }
             return p;
         } finally {
             em.close();
@@ -53,10 +56,13 @@ public class PersonFacade implements IFPersonFacade {
     }
 
     @Override
-    public Person deletePerson(int id) {
+    public Person deletePerson(int id) throws PersonNotFoundException {
         EntityManager em = getEntityManager();
         try {
             Person p = em.find(Person.class, id);
+            if (p == null) {
+                throw new PersonNotFoundException("No Person found with provided id");
+            }
             em.getTransaction().begin();
             em.remove(p);
             em.getTransaction().commit();
